Change collection properties back to original models
