+++ AdminWebsite/AdminWebsite/ClientApp/src/app/common/model/api-contract-to-client-model-mappers.ts
@@ -1,38 +1,29 @@
 import {
-    AvailableLanguageResponse,
     CaseResponse,
-    EndpointResponse,
     HearingDetailsResponse,
-    InterprepretationType,
     ParticipantResponse,
-    ScreeningResponse,
-    ScreeningType as ApiScreeningType,
-    JudiciaryParticipantResponse,
-    LinkedParticipantResponse
+    LinkedParticipantResponse,
+    EndpointResponse
 } from 'src/app/services/clients/api-client';
 import { Constants } from 'src/app/common/constants';
-import { v4 as uuid } from 'uuid';
-import {
-    VHBooking,
-    VHCase,
-    VHParticipant,
-    VHLinkedParticipant,
-    VHJudiciaryParticipant,
-    VHJudicaryRoleCode,
-    VHEndpoint,
-    VHInterpreterSelected,
-    VHScreening
-} from './vh-booking';
+import { VHBooking } from './vh-booking';
+import { CaseModel } from './case.model';
+import { InterpreterSelectedDto } from 'src/app/booking/interpreter-form/interpreter-selected.model';
+import { mapScreeningResponseToScreeningDto } from 'src/app/booking/screening/screening.model';
+import { ParticipantModel } from './participant.model';
+import { LinkedParticipantModel } from './linked-participant.model';
+import { JudicialMemberDto } from 'src/app/booking/judicial-office-holders/models/add-judicial-member.model';
+import { EndpointModel } from './endpoint.model';
 
 export function mapHearingToVHBooking(hearing: HearingDetailsResponse): VHBooking {
     return {
         hearing_id: hearing.id,
         scheduled_date_time: new Date(hearing.scheduled_date_time),
         scheduled_duration: hearing.scheduled_duration,
-        case: mapCaseToVHCase(hearing.cases[0]),
-        participants: hearing.participants?.map(participant => mapParticipantToVHParticipant(participant)),
+        case: mapCaseResponseToCaseModel(hearing.cases)[0],
+        participants: mapParticipantResponseToParticipantModel(hearing.participants),
         judiciaryParticipants: hearing.judiciary_participants?.map(judiciaryParticipant =>
-            mapJudiciaryParticipantToVHJudiciaryParticipant(judiciaryParticipant)
+            JudicialMemberDto.fromJudiciaryParticipantResponse(judiciaryParticipant)
         ),
         created_by: hearing.created_by,
         case_type: hearing.case_type_name,
@@ -46,7 +37,7 @@ export function mapHearingToVHBooking(hearing: HearingDetailsResponse): VHBookin
         updated_date: new Date(hearing.updated_date),
         status: hearing.status,
         audio_recording_required: hearing.audio_recording_required,
-        endpoints: hearing.endpoints?.map(endpoint => mapEndpointToVHEndpoint(endpoint, hearing.participants)),
+        endpoints: mapEndpointResponseToEndpointModel(hearing.endpoints, hearing.participants),
         isMultiDay: hearing.group_id !== null,
         multiDayHearingLastDayScheduledDateTime: hearing.multi_day_hearing_last_day_scheduled_date_time,
         hearingsInGroup: hearing.hearings_in_group?.map(hearingInGroup => mapHearingToVHBooking(hearingInGroup)),
@@ -55,120 +46,86 @@ export function mapHearingToVHBooking(hearing: HearingDetailsResponse): VHBookin
     };
 }
 
-export function mapCaseToVHCase(caseResponse: CaseResponse): VHCase {
-    return {
-        number: caseResponse.number,
-        name: caseResponse.name,
-        isLeadCase: caseResponse.is_lead_case
-    };
-}
-
-export function mapParticipantToVHParticipant(participant: ParticipantResponse): VHParticipant {
-    return {
-        id: participant.id,
-        externalReferenceId: participant.external_reference_id,
-        title: participant.title,
-        first_name: participant.first_name,
-        last_name: participant.last_name,
-        middle_names: participant.middle_names,
-        display_name: participant.display_name,
-        username: participant.username,
-        email: participant.contact_email,
-        hearing_role_name: participant.hearing_role_name,
-        hearing_role_code: participant.hearing_role_code,
-        phone: participant.telephone_number,
-        representee: participant.representee,
-        company: participant.organisation,
-        is_judge: participant.user_role_name === Constants.UserRoles.Judge,
-        linked_participants: participant.linked_participants?.map(linkedParticipant =>
-            mapLinkedParticipantResponseToVHLinkedParticipant(linkedParticipant)
-        ),
-        user_role_name: participant.user_role_name,
-        is_staff_member: participant.user_role_name === Constants.UserRoles.StaffMember,
-        interpretation_language: mapAvailableLanguageToVHInterpreterSelected(participant.interpreter_language),
-        screening: mapScreeningResponseToVHScreening(participant.screening_requirement)
-    };
-}
-
-export function mapLinkedParticipantResponseToVHLinkedParticipant(linkedParticipant: LinkedParticipantResponse): VHLinkedParticipant {
-    return {
-        linkType: linkedParticipant.type,
-        linkedParticipantId: linkedParticipant.linked_id
-    };
-}
-
-export function mapJudiciaryParticipantToVHJudiciaryParticipant(
-    judiciaryParticipant: JudiciaryParticipantResponse
-): VHJudiciaryParticipant {
-    return {
-        firstName: judiciaryParticipant.first_name,
-        lastName: judiciaryParticipant.last_name,
-        fullName: judiciaryParticipant.full_name,
-        email: judiciaryParticipant.email,
-        telephone: judiciaryParticipant.work_phone,
-        personalCode: judiciaryParticipant.personal_code,
-        isGeneric: judiciaryParticipant.is_generic,
-        optionalContactNumber: judiciaryParticipant.optional_contact_telephone,
-        optionalContactEmail: judiciaryParticipant.optional_contact_email,
-        roleCode: judiciaryParticipant.role_code as VHJudicaryRoleCode,
-        displayName: judiciaryParticipant.display_name,
-        interpretationLanguage: mapAvailableLanguageToVHInterpreterSelected(judiciaryParticipant.interpreter_language)
-    };
-}
-
-export function mapEndpointToVHEndpoint(endpoint: EndpointResponse, participants: ParticipantResponse[]): VHEndpoint {
-    const defenceAdvocate = participants.find(p => p.id === endpoint.defence_advocate_id);
-    return {
-        externalReferenceId: endpoint.external_reference_id || uuid(),
-        id: endpoint.id,
-        displayName: endpoint.display_name,
-        sip: endpoint.sip,
-        pin: endpoint.pin,
-        defenceAdvocate: defenceAdvocate?.contact_email,
-        interpretationLanguage: mapAvailableLanguageToVHInterpreterSelected(endpoint.interpreter_language),
-        screening: mapScreeningResponseToVHScreening(endpoint.screening_requirement)
-    };
-}
-
-export function mapAvailableLanguageToVHInterpreterSelected(availableLanguage: AvailableLanguageResponse): VHInterpreterSelected {
-    if (!availableLanguage) {
-        return null;
+export function mapCaseResponseToCaseModel(casesResponse: CaseResponse[]): CaseModel[] {
+    const cases: CaseModel[] = [];
+    let caseRequest: CaseModel;
+    if (casesResponse && casesResponse.length > 0) {
+        casesResponse.forEach(c => {
+            caseRequest = new CaseModel();
+            caseRequest.name = c.name;
+            caseRequest.number = c.number;
+            caseRequest.isLeadCase = c.is_lead_case;
+            cases.push(caseRequest);
+        });
     }
-
-    const interpreterSelected: VHInterpreterSelected = {
-        interpreterRequired: true,
-        spokenLanguageCode: null,
-        spokenLanguageCodeDescription: null,
-        signLanguageCode: null,
-        signLanguageDescription: null
-    };
-    switch (availableLanguage.type) {
-        case InterprepretationType.Verbal:
-            interpreterSelected.spokenLanguageCode = availableLanguage.code;
-            interpreterSelected.spokenLanguageCodeDescription = availableLanguage.description;
-            break;
-        case InterprepretationType.Sign:
-            interpreterSelected.signLanguageCode = availableLanguage.code;
-            interpreterSelected.signLanguageDescription = availableLanguage.description;
-            break;
-        default:
-            throw new Error(`Unknown interpretation type ${availableLanguage.type}`);
-    }
-
-    return interpreterSelected;
+    return cases;
 }
 
-export function mapScreeningResponseToVHScreening(response: ScreeningResponse): VHScreening {
-    if (!response) {
-        return undefined;
+export function mapParticipantResponseToParticipantModel(response: ParticipantResponse[]): ParticipantModel[] {
+    const participants: ParticipantModel[] = [];
+    let participant: ParticipantModel;
+    if (response && response.length > 0) {
+        response.forEach(p => {
+            participant = new ParticipantModel();
+            participant.id = p.id;
+            participant.title = p.title;
+            participant.first_name = p.first_name;
+            participant.middle_names = p.middle_names;
+            participant.last_name = p.last_name;
+            participant.username = p.username;
+            participant.display_name = p.display_name;
+            participant.email = p.contact_email;
+            participant.phone = p.telephone_number;
+            participant.hearing_role_name = p.hearing_role_name;
+            participant.hearing_role_code = p.hearing_role_code;
+            participant.representee = p.representee;
+            participant.company = p.organisation;
+            participant.is_judge = p.user_role_name === Constants.UserRoles.Judge;
+            participant.is_staff_member = p.user_role_name === Constants.UserRoles.StaffMember;
+            participant.linked_participants = mapLinkedParticipantResponseToLinkedParticipantModel(p.linked_participants);
+            participant.user_role_name = p.user_role_name;
+            participant.interpretation_language = InterpreterSelectedDto.fromAvailableLanguageResponse(p.interpreter_language);
+            participant.screening = mapScreeningResponseToScreeningDto(p.screening_requirement);
+            if (p.external_reference_id) {
+                // only override the external reference id if it is not null else ParticipantModel will initialise to a UUID in the ctor
+                participant.externalReferenceId = p.external_reference_id;
+            }
+            participants.push(participant);
+        });
     }
+    return participants;
+}
 
-    const mappedProtectFrom = response.protect_from.map(protectFrom => ({
-        externalReferenceId: protectFrom
-    }));
+export function mapLinkedParticipantResponseToLinkedParticipantModel(response: LinkedParticipantResponse[]): LinkedParticipantModel[] {
+    const linkedParticipants: LinkedParticipantModel[] = [];
+    let linkedParticipant: LinkedParticipantModel;
+    if (response && response.length > 0) {
+        response.forEach(p => {
+            linkedParticipant = new LinkedParticipantModel();
+            linkedParticipant.linkType = p.type;
+            linkedParticipant.linkedParticipantId = p.linked_id;
+            linkedParticipants.push(linkedParticipant);
+        });
+    }
+    return linkedParticipants;
+}
 
-    return {
-        measureType: response.type === ApiScreeningType.All ? 'All' : 'Specific',
-        protectFrom: mappedProtectFrom
-    };
+export function mapEndpointResponseToEndpointModel(response: EndpointResponse[], participants: ParticipantResponse[]): EndpointModel[] {
+    const endpoints: EndpointModel[] = [];
+    let endpoint: EndpointModel;
+    if (response && response.length > 0) {
+        response.forEach(e => {
+            const defenceAdvocate = participants.find(p => p.id === e.defence_advocate_id);
+            endpoint = new EndpointModel(e.external_reference_id);
+            endpoint.id = e.id;
+            endpoint.displayName = e.display_name;
+            endpoint.pin = e.pin;
+            endpoint.sip = e.sip;
+            endpoint.defenceAdvocate = defenceAdvocate?.contact_email;
+            endpoint.interpretationLanguage = InterpreterSelectedDto.fromAvailableLanguageResponse(e.interpreter_language);
+            endpoint.screening = mapScreeningResponseToScreeningDto(e.screening_requirement);
+            endpoints.push(endpoint);
+        });
+    }
+    return endpoints;
 }
