+++ AdminWebsite/AdminWebsite/ClientApp/src/app/common/model/vh-booking.ts
@@ -1,12 +1,17 @@
-import { LinkedParticipantType, VideoSupplier } from 'src/app/services/clients/api-client';
+import { VideoSupplier } from 'src/app/services/clients/api-client';
+import { CaseModel } from './case.model';
+import { ParticipantModel } from './participant.model';
+import { JudicialMemberDto } from 'src/app/booking/judicial-office-holders/models/add-judicial-member.model';
+import { EndpointModel } from './endpoint.model';
+import { LinkedParticipantModel } from './linked-participant.model';
 
 export interface VHBooking {
     hearing_id?: string;
     scheduled_date_time?: Date;
     scheduled_duration?: number;
-    case?: VHCase;
-    participants?: VHParticipant[];
-    judiciaryParticipants?: VHJudiciaryParticipant[];
+    case?: CaseModel;
+    participants?: ParticipantModel[];
+    judiciaryParticipants?: JudicialMemberDto[];
     created_by?: string;
     case_type?: string;
     case_type_service_id?: string;
@@ -21,11 +26,11 @@ export interface VHBooking {
     updated_date: Date;
     status?: string;
     audio_recording_required?: boolean;
-    endpoints?: VHEndpoint[];
+    endpoints?: EndpointModel[];
     isMultiDayEdit?: boolean;
     end_hearing_date_time?: Date;
     telephone_conference_id?: string;
-    linked_participants?: VHLinkedParticipant[];
+    linked_participants?: LinkedParticipantModel[];
     hearing_dates?: Date[];
     isConfirmed?: boolean;
     isMultiDay?: boolean;
@@ -47,108 +52,4 @@ export function createVHBooking(): VHBooking {
         updated_date: new Date(),
         supplier: VideoSupplier.Vodafone
     };
-}
-
-export interface VHCase {
-    number?: string;
-    name?: string;
-    isLeadCase?: boolean;
-}
-
-export interface VHParticipant {
-    id?: string;
-    externalReferenceId?: string;
-    title?: string;
-    first_name?: string;
-    last_name?: string;
-    middle_names?: string;
-    display_name?: string;
-    username?: string;
-    email?: string;
-    hearing_role_name?: string;
-    hearing_role_code?: string;
-    phone?: string;
-    representee?: string;
-    company?: string;
-    is_judge?: boolean;
-    is_exist_person?: boolean;
-    interpreterFor?: string;
-    linked_participants?: VHLinkedParticipant[];
-    interpretee_name?: string;
-    is_interpretee?: boolean;
-    user_role_name?: string;
-    is_courtroom_account?: boolean;
-    addedDuringHearing?: boolean;
-    is_staff_member?: boolean;
-    contact_email?: string;
-    isJudiciaryMember?: boolean;
-    interpretation_language: VHInterpreterSelected;
-    screening?: VHScreening;
-}
-
-export interface VHLinkedParticipant {
-    id?: string;
-    participantId?: string;
-    participantEmail?: string;
-    linkedParticipantId?: string;
-    linkedParticipantEmail?: string;
-    linkType?: LinkedParticipantType;
-}
-
-export interface VHJudiciaryParticipant {
-    firstName: string;
-    lastName: string;
-    fullName: string;
-    email: string;
-    telephone: string;
-    personalCode: string;
-    isGeneric: boolean;
-    roleCode?: VHJudicaryRoleCode;
-    displayName: string;
-    optionalContactNumber: string;
-    optionalContactEmail: string;
-    interpretationLanguage: VHInterpreterSelected;
-}
-
-export type VHJudicaryRoleCode = 'Judge' | 'PanelMember';
-
-export interface VHEndpoint {
-    externalReferenceId: string;
-    id?: string;
-    displayName?: string;
-    sip?: string;
-    pin?: string;
-    /**
-     * Defence advocate email address, not their ID
-     */
-    defenceAdvocate?: string;
-    username?: string;
-    contactEmail?: string;
-    interpretationLanguage: VHInterpreterSelected;
-    screening?: VHScreening;
-}
-
-export interface VHVenue {
-    id: number;
-    name: string;
-    code: string;
-}
-
-export interface VHInterpreterSelected {
-    interpreterRequired: boolean;
-    signLanguageCode?: string;
-    signLanguageDescription?: string;
-    spokenLanguageCode?: string;
-    spokenLanguageCodeDescription?: string;
-}
-
-export interface VHScreening {
-    measureType: ScreeningType;
-    protectFrom: VHProtectFrom[];
-}
-
-export type ScreeningType = 'All' | 'Specific';
-
-export interface VHProtectFrom {
-    externalReferenceId: string;
 }
