+++ AdminWebsite/AdminWebsite/ClientApp/src/app/services/video-hearings.service.ts
@@ -7,14 +7,12 @@ import {
     BookingStatus,
     CancelMultiDayHearingRequest,
     CaseRequest,
-    CaseResponse,
     EditCaseRequest,
     EditEndpointRequest,
     EditHearingRequest,
     EditMultiDayHearingRequest,
     EditParticipantRequest,
     EndpointRequest,
-    EndpointResponse,
     HearingDetailsResponse,
     HearingRoleResponse,
     HearingTypeResponse,
@@ -22,16 +20,13 @@ import {
     JudiciaryParticipantRequest,
     LinkedParticipant,
     LinkedParticipantRequest,
-    LinkedParticipantResponse,
     MultiHearingRequest,
     ParticipantRequest,
-    ParticipantResponse,
     PhoneConferenceResponse,
     SpecialMeasureScreeningRequest,
     UpdateBookingStatusResponse,
     UpdateHearingInGroupRequest
 } from './clients/api-client';
-import { CaseModel } from '../common/model/case.model';
 import { ParticipantModel } from '../common/model/participant.model';
 import { EndpointModel } from '../common/model/endpoint.model';
 import { LinkedParticipantModel } from '../common/model/linked-participant.model';
@@ -40,7 +35,7 @@ import * as moment from 'moment';
 import { JudicialMemberDto } from '../booking/judicial-office-holders/models/add-judicial-member.model';
 import { map, shareReplay } from 'rxjs/operators';
 import { InterpreterSelectedDto } from '../booking/interpreter-form/interpreter-selected.model';
-import { mapScreeningResponseToScreeningDto, ScreeningDto } from '../booking/screening/screening.model';
+import { ScreeningDto } from '../booking/screening/screening.model';
 import { ReferenceDataService } from './reference-data.service';
 import { createVHBooking, VHBooking } from '../common/model/vh-booking';
 import { mapHearingToVHBooking } from '../common/model/api-contract-to-client-model-mappers';
@@ -353,21 +348,6 @@ export class VideoHearingsService {
         return cases;
     }
 
-    mapCaseResponseToCaseModel(casesResponse: CaseResponse[]): CaseModel[] {
-        const cases: CaseModel[] = [];
-        let caseRequest: CaseModel;
-        if (casesResponse && casesResponse.length > 0) {
-            casesResponse.forEach(c => {
-                caseRequest = new CaseModel();
-                caseRequest.name = c.name;
-                caseRequest.number = c.number;
-                caseRequest.isLeadCase = c.is_lead_case;
-                cases.push(caseRequest);
-            });
-        }
-        return cases;
-    }
-
     mapJudicialMemberDtoToJudiciaryParticipantRequest(judicialMemberDtos: JudicialMemberDto[]): JudiciaryParticipantRequest[] {
         return judicialMemberDtos.map(judicialMemberDto => {
             const judiciaryParticipantRequest: JudiciaryParticipantRequest = new JudiciaryParticipantRequest({
@@ -436,75 +416,6 @@ export class VideoHearingsService {
             screen_all: false,
             screen_from_external_reference_ids: screeningDto.protectFrom.map(x => x.externalReferenceId)
         });
-    }
-
-    mapParticipantResponseToParticipantModel(response: ParticipantResponse[]): ParticipantModel[] {
-        const participants: ParticipantModel[] = [];
-        let participant: ParticipantModel;
-        if (response && response.length > 0) {
-            response.forEach(p => {
-                participant = new ParticipantModel();
-                participant.id = p.id;
-                participant.title = p.title;
-                participant.first_name = p.first_name;
-                participant.middle_names = p.middle_names;
-                participant.last_name = p.last_name;
-                participant.username = p.username;
-                participant.display_name = p.display_name;
-                participant.email = p.contact_email;
-                participant.phone = p.telephone_number;
-                participant.hearing_role_name = p.hearing_role_name;
-                participant.hearing_role_code = p.hearing_role_code;
-                participant.representee = p.representee;
-                participant.company = p.organisation;
-                participant.is_judge = p.user_role_name === Constants.UserRoles.Judge;
-                participant.is_staff_member = p.user_role_name === Constants.UserRoles.StaffMember;
-                participant.linked_participants = this.mapLinkedParticipantResponseToLinkedParticipantModel(p.linked_participants);
-                participant.user_role_name = p.user_role_name;
-                participant.interpretation_language = InterpreterSelectedDto.fromAvailableLanguageResponse(p.interpreter_language);
-                participant.screening = mapScreeningResponseToScreeningDto(p.screening_requirement);
-                if (p.external_reference_id) {
-                    // only override the external reference id if it is not null else ParticipantModel will initialise to a UUID in the ctor
-                    participant.externalReferenceId = p.external_reference_id;
-                }
-                participants.push(participant);
-            });
-        }
-        return participants;
-    }
-
-    mapLinkedParticipantResponseToLinkedParticipantModel(response: LinkedParticipantResponse[]): LinkedParticipantModel[] {
-        const linkedParticipants: LinkedParticipantModel[] = [];
-        let linkedParticipant: LinkedParticipantModel;
-        if (response && response.length > 0) {
-            response.forEach(p => {
-                linkedParticipant = new LinkedParticipantModel();
-                linkedParticipant.linkType = p.type;
-                linkedParticipant.linkedParticipantId = p.linked_id;
-                linkedParticipants.push(linkedParticipant);
-            });
-        }
-        return linkedParticipants;
-    }
-
-    mapEndpointResponseToEndpointModel(response: EndpointResponse[], participants: ParticipantResponse[]): EndpointModel[] {
-        const endpoints: EndpointModel[] = [];
-        let endpoint: EndpointModel;
-        if (response && response.length > 0) {
-            response.forEach(e => {
-                const defenceAdvocate = participants.find(p => p.id === e.defence_advocate_id);
-                endpoint = new EndpointModel(e.external_reference_id);
-                endpoint.id = e.id;
-                endpoint.displayName = e.display_name;
-                endpoint.pin = e.pin;
-                endpoint.sip = e.sip;
-                endpoint.defenceAdvocate = defenceAdvocate?.contact_email;
-                endpoint.interpretationLanguage = InterpreterSelectedDto.fromAvailableLanguageResponse(e.interpreter_language);
-                endpoint.screening = mapScreeningResponseToScreeningDto(e.screening_requirement);
-                endpoints.push(endpoint);
-            });
-        }
-        return endpoints;
     }
 
     mapLinkedParticipants(linkedParticipantModels: LinkedParticipantModel[] = []): LinkedParticipantRequest[] {
